backend/enhance: add time based filter for driver pool config
