+++ Backend/app/dashboard/provider-dashboard/package.yaml
@@ -111,6 +111,7 @@ library:
           - -fmax-simplifier-iterations1
           - -fno-specialise-aggressively
           - -j4
+          - -fwhole-archive-hs-libs
       else:
         ghc-options:
           - -O2
@@ -143,6 +144,7 @@ executables:
           - -fmax-simplifier-iterations1
           - -fno-specialise-aggressively
           - -j4
+          - -fwhole-archive-hs-libs
       else:
         ghc-options:
           - -O2
