+++ Backend/app/dashboard/provider-dashboard/provider-dashboard.cabal
@@ -128,7 +128,7 @@ library
     , utf8-string
   default-language: Haskell2010
   if flag(Local)
-    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4
+    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4 -fwhole-archive-hs-libs
   else
     ghc-options: -O2 -j4
 
@@ -204,6 +204,6 @@ executable provider-dashboard-exe
     , utf8-string
   default-language: Haskell2010
   if flag(Local)
-    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4
+    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4 -fwhole-archive-hs-libs
   else
     ghc-options: -O2 -j4
