+++ Backend/app/dashboard/rider-dashboard/package.yaml
@@ -104,6 +104,7 @@ library:
           - -fno-cse
           - -fmax-simplifier-iterations1
           - -fno-specialise-aggressively
+          - -fwhole-archive-hs-libs
           - -j4
       else:
         ghc-options:
@@ -133,6 +134,7 @@ executables:
           - -fno-cse
           - -fmax-simplifier-iterations1
           - -fno-specialise-aggressively
+          - -fwhole-archive-hs-libs
           - -j4
       else:
         ghc-options:
