+++ Backend/app/dashboard/rider-dashboard/rider-dashboard.cabal
@@ -118,7 +118,7 @@ library
     , unordered-containers
   default-language: Haskell2010
   if flag(Local)
-    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4
+    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -fwhole-archive-hs-libs -j4
   else
     ghc-options: -O2 -j4
 
@@ -188,6 +188,6 @@ executable rider-dashboard-exe
     , unordered-containers
   default-language: Haskell2010
   if flag(Local)
-    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4
+    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -fwhole-archive-hs-libs -j4
   else
     ghc-options: -O2 -j4
