+++ Backend/app/provider-platform/dynamic-offer-driver-app/Allocator/driver-offer-allocator.cabal
@@ -112,7 +112,7 @@ library
     , warp
   default-language: Haskell2010
   if flag(Local)
-    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4
+    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4 -fwhole-archive-hs-libs
   else
     ghc-options: -O2 -j4
 
@@ -199,6 +199,6 @@ executable driver-offer-allocator-exe
     , warp
   default-language: Haskell2010
   if flag(Local)
-    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4
+    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4 -fwhole-archive-hs-libs
   else
     ghc-options: -O2 -j4
