+++ Backend/app/provider-platform/dynamic-offer-driver-app/Allocator/package.yaml
@@ -116,6 +116,7 @@ library:
           - -fmax-simplifier-iterations1
           - -fno-specialise-aggressively
           - -j4
+          - -fwhole-archive-hs-libs
       else:
         ghc-options:
           - -O2
@@ -148,6 +149,7 @@ executables:
             - -fmax-simplifier-iterations1
             - -fno-specialise-aggressively
             - -j4
+            - -fwhole-archive-hs-libs
         else:
           ghc-options:
             - -O2
