+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/package.yaml
@@ -148,6 +148,7 @@ library:
           - -fmax-simplifier-iterations1
           - -fno-specialise-aggressively
           - -j4
+          - -fwhole-archive-hs-libs
       else:
         ghc-options:
           - -O2
@@ -184,6 +185,7 @@ executables:
           - -fmax-simplifier-iterations1
           - -fno-specialise-aggressively
           - -j4
+          - -fwhole-archive-hs-libs
       else:
         ghc-options:
           - -O2
