+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src-read-only/Domain/Types/FareProduct.hs
@@ -1,18 +1,16 @@
 {-# LANGUAGE ApplicativeDo #-}
 {-# LANGUAGE TemplateHaskell #-}
-{-# OPTIONS_GHC -Wno-dodgy-exports #-}
 {-# OPTIONS_GHC -Wno-unused-imports #-}
 
-module Domain.Types.FareProduct (module Domain.Types.FareProduct, module ReExport) where
+module Domain.Types.FareProduct where
 
 import Data.Aeson
 import qualified Domain.Types.Common
-import Domain.Types.Extra.FareProduct as ReExport
-import qualified Domain.Types.Extra.FareProduct
 import qualified Domain.Types.FarePolicy
 import qualified Domain.Types.Merchant
 import qualified Domain.Types.MerchantOperatingCity
 import qualified Domain.Types.ServiceTierType
+import qualified Domain.Types.TimeBound
 import Kernel.Prelude
 import qualified Kernel.Types.Id
 import qualified Lib.Types.SpecialLocation
@@ -26,7 +24,7 @@ data FareProduct = FareProduct
     merchantId :: Kernel.Types.Id.Id Domain.Types.Merchant.Merchant,
     merchantOperatingCityId :: Kernel.Types.Id.Id Domain.Types.MerchantOperatingCity.MerchantOperatingCity,
     searchSource :: Domain.Types.FareProduct.SearchSource,
-    timeBounds :: Domain.Types.Extra.FareProduct.TimeBound,
+    timeBounds :: Domain.Types.TimeBound.TimeBound,
     tripCategory :: Domain.Types.Common.TripCategory,
     vehicleServiceTier :: Domain.Types.ServiceTierType.ServiceTierType
   }
