+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src-read-only/Storage/Queries/DriverPoolConfig.hs
@@ -63,6 +63,7 @@ updateByPrimaryKey (Domain.Types.DriverPoolConfig.DriverPoolConfig {..}) = do
       Se.Set Beam.scheduleTryTimes scheduleTryTimes,
       Se.Set Beam.singleBatchProcessTime singleBatchProcessTime,
       Se.Set Beam.thresholdToIgnoreActualDistanceThreshold thresholdToIgnoreActualDistanceThreshold,
+      Se.Set Beam.timeBounds timeBounds,
       Se.Set Beam.tripCategory tripCategory,
       Se.Set Beam.tripDistance tripDistance,
       Se.Set Beam.updatedAt _now,
@@ -106,6 +107,7 @@ instance FromTType' Beam.DriverPoolConfig Domain.Types.DriverPoolConfig.DriverPo
             scheduleTryTimes = scheduleTryTimes,
             singleBatchProcessTime = singleBatchProcessTime,
             thresholdToIgnoreActualDistanceThreshold = thresholdToIgnoreActualDistanceThreshold,
+            timeBounds = timeBounds,
             tripCategory = tripCategory,
             tripDistance = tripDistance,
             updatedAt = updatedAt,
@@ -146,6 +148,7 @@ instance ToTType' Beam.DriverPoolConfig Domain.Types.DriverPoolConfig.DriverPool
         Beam.scheduleTryTimes = scheduleTryTimes,
         Beam.singleBatchProcessTime = singleBatchProcessTime,
         Beam.thresholdToIgnoreActualDistanceThreshold = thresholdToIgnoreActualDistanceThreshold,
+        Beam.timeBounds = timeBounds,
         Beam.tripCategory = tripCategory,
         Beam.tripDistance = tripDistance,
         Beam.updatedAt = updatedAt,
