+++ Backend/app/safety-dashboard/package.yaml
@@ -120,6 +120,7 @@ library:
           - -fno-cse
           - -fmax-simplifier-iterations1
           - -fno-specialise-aggressively
+          - -fwhole-archive-hs-libs
           - -j4
       else:
         ghc-options:
@@ -148,6 +149,7 @@ executables:
           - -fno-cse
           - -fmax-simplifier-iterations1
           - -fno-specialise-aggressively
+          - -fwhole-archive-hs-libs
           - -j4
       else:
         ghc-options:
