+++ Backend/app/safety-dashboard/safety-dashboard.cabal
@@ -165,7 +165,7 @@ library
     , unordered-containers
   default-language: Haskell2010
   if flag(Local)
-    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4
+    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -fwhole-archive-hs-libs -j4
   else
     ghc-options: -O2 -j4
 
@@ -246,6 +246,6 @@ executable safety-dashboard-exe
     , unordered-containers
   default-language: Haskell2010
   if flag(Local)
-    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -j4
+    ghc-options: -O0 -funfolding-use-threshold20 -fno-cse -fmax-simplifier-iterations1 -fno-specialise-aggressively -fwhole-archive-hs-libs -j4
   else
     ghc-options: -O2 -j4
